Extract the SwipeDownBar out of the NowPlaying widget
